Pathfinding, models & soundtrack updates.
Updated the pathfinding integration with a more optimized BFS algorithm that utilizes both a queue & a stack for more efficiency & responsiveness.
Updated the spawn-location algorithm for the AI for better performance & spawn locations to avoid cluster.
Updated the in game models of the walls and borders (sharper/more depth)
Added a soundtrack during gameplay, death, and winning.
- Finished the project, farewell scrayville.